Made module compatible with node v6

+++ performSetup.js
@@ -11,15 +11,17 @@ var tmpFeatureFiles = new featureFileSplitter();
  * @param {string} [options.tagExpression] - Tag expression to parse
  * @param {string} [options.ff] - Feature File Name to parse
  * @param {string} [options.lang] - Language of sourceSpecDirectory
- *
+ * @param {Boolean} [options.cleanTmpSpecDirectory] - Boolean for cleaning the Temp Spec Directory 
  * @return {Promise<void>}
  */
 var performSetup = function performSetup(options) {
     try {
 
-        //Remove Tmp Spec Directory & Create One
-        fs_extra.removeSync(options.tmpSpecDirectory);
-        fs_extra.ensureDirSync(options.tmpSpecDirectory);
+        if (options.cleanTmpSpecDirectory) {
+            //Remove Tmp Spec Directory during setup & Create One
+            fs_extra.removeSync(options.tmpSpecDirectory);
+            fs_extra.ensureDirSync(options.tmpSpecDirectory);
+        }
 
         //Compile and Create Split Feature Files
         tmpFeatureFiles.compile({
